progress towards command overlays
Initial steps towards overlaying the gripper and navigation
videos with clickable command regions. For the navigation mode,
this commit provides an unrefined overlay to drive the mobile base
and an unrefined overlay to operate the gripper and wrist. No
overlays are available to move the lift and the telescoping arm.
Also, the overlays don't change when the videos change due to
activation of manipulation mode.