fully functional gripper & navigation camera interface
This is a fully functional version of the interface that uses a
wide-angle gripper camera and a wide-angle navigation camera.

With this interface, I was able to pick up a computer mouse
and carry it while navigating.

The interface expects the wide-angle gripper camera to be
attached as far down the gripper as it can be (i.e., as close
to the fingers as it can be). It expects the navigation camera
to be attached to the top of the head and positioned so that it is
pointing straight down at the ground and looking at the
center front of the mobile base. The top of the navigation
camera should be pointed in the direction that the telescoping arm
extends.

The gripper camera should be plugged into the wrist USB port,
and the navigation camera should be plugged into the head USB port.
Nothing else should be plugged into these ports, since bandwidth
is limited and the udev rules use the USB port topology to find
and distinguish the gripper and navigation cameras.

Limitations of this version include the following:

 + The video is low resolution. 320x240 for each camera due to
   USB bandwidth issues. 640x480 would be much better. Hardware
   changes might make this possible in the future, but software
   is unlikely to improve the resolution with the current cameras.

 + The video is a little distorted. It is slightly elongated in
   one direction for each camera. This distortion doesn't appear in
   640x480 mode. It seems to be an issue with the 320x240 mode for
   these USB cameras. This could be corrected with software,
   but I'm hopeful that a future resolution improvement will
   resolve this.

 + The interface video has large unused black regions. The
   individual gripper and navigation videos could be cropped or
   otherwise transformed to make better use of the video display.