high resolution video & general cleanup
The wide-angle camera web interface now uses high-resolution
video at a low frame rate. Both cameras worked together at
1024x768 resolution and 6 frames per second. Trading resolution
for frames per second appears to be worthwhile, since the robot
moves slowly and haptic feedback (joint torque visualizations)
is updated frequently and has a low latency, which helps
compensate for the low video rate during teleoperation.

This update also streamlines and cleans up the code and branch
via changes like the following:

  + specializes this branch for the wide-angle camera web interface
     + removes D435i (Stretch's head camera) throughout the branch
     + removes individual camera launch files
     + removes individual camera installation files
     + provides a single launch file for the interface
  + no longer displays video streams on the robot using ROS (they
    can still be viewed in the robot's browser)

One remaining issue is that the default camera settings did not
work as expected, such as the gain settings. The video streams
were unusually dark and difficult to use. This update explicitly
attempts to set the camera parameters when launching the cameras
in ROS with usb_cam.

There are at least two potential issues with this approach.
First, setting the parameters via usb_cam may have problems.
It did not seem as clear and reliable as expected. Second,
the automatic camera parameters are turned off in an attempt
to avoid issues. Consequently, the cameras might not behave
well in different lighting conditions, since they no longer
have the ability to adjust automatically.

To test various parameters and interrogate the cameras, one
can install and use guvcview.

To install this GUI application, use the following command:

   sudo apt install guvcview

Now, you can go to the "Show Applications" menu, search for
guvcview, run the app, and connect it to one of the USB 2.0
cameras. This should not be run at the same time as the web
interface and web interface ROS code. Once guvcview is running,
you can use it to see video from the camera while adjusting the
camera's settings. You can also adjust the parameters in the
cameras' launch files:

   ./launch/gripper_camera.launch
   ./launch/navigation_camera.launch