Turn Content Security Policy (CSP) off by default
While better for security, having the Content Security Policy (CSP) active by default has resulted in challenges for other developers and possibly interfered with use. This should be enabled for deployed systems on the Internet. Having it deactivated is only appropriate for a secure local area network (LAN) with a strong firewall.